requirements.txt
pytest~=8.3.2
selenium~=4.25.0
allure-pytest~=2.13.5
allure-python-commons==2.13.5
webdriver-manager
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
requests~=2.32.3